for a pride mode option that will be in options
gay + asexual + aromantic flags will be added soon! 
pride mode option will change the title screen to be a random lgbtq+ flag (specific ones may appear more frequently on certain days or months like the lgbtqlogotemp.jpg appearing more on pride month, bisexuallogotemp.jpg appearing more on bisexual visibility month (that's this month!!!) and translogotemp.jpg appearing more on transgender visibility day

if you're wondering why this game would need this, the answer is that this game does not need this, but i wanna add it anyways. perhaps it could be in a special build of the game although i wanna include it in the main game.